# БР1.2/Москалец Данила/homeworks/hw2/openapi.yaml
openapi: 3.0.0
info:
  title: Sangel API
  description: Маркетплейс охранных услуг
  version: 1.0.0
  contact:
    name: Москалец Данила Алексеевич
servers:
  - url: https://api.sangel.ru/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        has_more:
          type: boolean

    # User
...
    Request:
      type: object
      properties:
        id:
          type: integer
        service:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            company:
              type: object
              properties:
                id:
                  type: integer
                title:
                  type: string
        user:
          type: object
          properties:
            id:
              type: integer
            first_name:
              type: string
            last_name:
              type: string
        status:
          type: string
          enum: [PENDING, ACCEPTED, REJECTED, CANCELLED]
        description:
          type: string
        reply:
          type: string
        created_at:
          type: string
          format: date-time
    CreateRequestRequest:
      type: object
      properties:
        description:
          type: string
    UpdateStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [ACCEPTED, REJECTED]
        reply:
          type: string

    # Review
    Review:
      type: object
      properties:
        id:
          type: integer
        service_id:
          type: integer
        service_name:
          type: string
        company_id:
          type: integer
        company_title:
          type: string
        user:
          type: object
          properties:
            id:
              type: integer
            first_name:
              type: string
            last_name:
              type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        created_at:
          type: string
          format: date-time
    CreateReviewRequest:
      type: object
      required:
        - rating
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    # Favorite
...
security:
  - bearerAuth: []
paths:
...
  /companies/{company_id}/requests:
    get:
      summary: Входящие заявки компании
      tags: [Requests]
      parameters:
        - name: company_id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, ACCEPTED, REJECTED, CANCELLED]
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Request'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        403:
          description: Доступ запрещен
        404:
          description: Компания не найдена
  /services/{service_id}/create_requests:
    post:
      summary: Создать заявку
      tags: [Requests]
      parameters:
        - name: service_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestRequest'
      responses:
        201:
          description: Заявка создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Request'
        403:
          description: Нельзя создать заявку на свою услугу
        404:
          description: Услуга не найдена
  /requests/{id}/status:
    put:
      summary: Изменить статус заявки
      tags: [Requests]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        200:
          description: Статус изменен
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Request'
        400:
          description: Некорректный статус
        403:
          description: Доступ запрещен
        404:
          description: Заявка не найдена
  /requests/{id}/cancel:
    put:
      summary: Отменить заявку
      tags: [Requests]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Заявка отменена
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Request'
        400:
          description: Невозможно отменить заявку
        403:
          description: Доступ запрещен
        404:
          description: Заявка не найдена

  # ==================== REVIEWS ====================
  /services/{service_id}/reviews:
    get:
      summary: Отзывы об услуге
      tags: [Reviews]
      security: []
      parameters:
        - name: service_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'

    post:
      summary: Оставить отзыв
      tags: [Reviews]
      parameters:
        - name: service_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        201:
          description: Отзыв оставлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Review'
        403:
          description: Нельзя оставить отзыв
        404:
          description: Услуга не найдена
  /reviews/{id}:
    delete:
      summary: Удалить отзыв
      tags: [Reviews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Отзыв удален
        403:
          description: Доступ запрещен
        404:
          description: Отзыв не найден

  # ==================== FAVORITES ====================
...